refactor: domain으로 코드 이동
