+++ src/main/java/com/example/weather/domain/location/repository/LocationRepository.java
@@ -1,7 +1,7 @@
-package com.example.weather.repository;
+package com.example.weather.domain.location.repository;
 
-import com.example.weather.domain.Location;
-import com.example.weather.domain.User;
+import com.example.weather.domain.location.entity.Location;
+import com.example.weather.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
