+++ src/main/java/com/example/weather/domain/user/entity/User.java
@@ -1,4 +1,4 @@
-package com.example.weather.domain;
+package com.example.weather.domain.user.entity;
 
 import jakarta.persistence.*;
 import java.time.LocalDateTime;
